did some refactoring for seeding script
